Removed accursed for loop from previous commit
Realized another obvious way to do the favorite faction
calculations was right in front of me. This patch removes
the nested for loop from before and instead includes the faction
in the game data used in the next part of the script.

Tested locally, runs in about the same amount of time as live.