Hotfixing
Minimal changes to see if it works in the test server

Added index.js to default.pug since the navigation bar currently has its js in that file, it needs to have its own js which will be added. This is being done since currently the js works on local machine but doesn't work on the test server.