Landing Page Update
# PUG
## index.pug

The multiple and slide variants/containers were taken out and replaced by the improvement container. The latter is cleaner and easier to read, and takes less user imput for the same output.

The rectangle containers were improved so they take less space and look somewhat cleaner.

Other unused classes/code was cleaned for better readability.
## default.pug
a link was added to the FAF logo on the footer, favicon image was also added.

# SASS
## _navigation.sass

Font sizes were all decreased so the website is easier to read and less on the "face" of the user.

FAF logo on the footer is now aligned with the other footer elements.

## index.sass
Creation of the styling of the improvement container, made the thinContainer titles have a z-index so the splat doesn't overlap them. Made other images and items smaller for better readability.

# Images
Some images were changed/updated to have a correct size ration for the rectangle container (from 1:1 to 16:9).