$Variables
# PUG
### ai.pug
Made the rectangle containers more compact (again, less on the "face" of the reader).

### default.pug
Made a comment in a nav_absolute that seems misplaced but isn't.

# SASS
### _navigation.sass and index.sass

Starting to update most of the css variable var(--) for the better $variable used by sass. This is being done so @Mixins can be used soon.

### mixins.sass
Cleared out of unused Mixins and updated to become the current variable.sass. It doesn't have any mixins but depending on the amount of mixins used, I might just not make a different page for variables and one for mixins (less pages = better).

### _mixins.sass and _footer.sass
Both deleted as they are out of use.

# JS
### index.js
Made extra comments so JS is even clearer and updated the pulsating function to work without var(--) (JS can't use $color because it will render $color into the css rather than the actual color variable. Ergo, manual colors were used).

#Images
### footerbanner.jpg
Updated to a better version of it that will be better displayed on smaller screens.