Re-structuring + Cleanup
## Deletion
Files deleted were 99% related to the styling of the website. Due to the new development of the website, these style sheets don't work anymore unless I stick to the naming conventions and previous work done. However, a great part of these files either seem to bear no purpose or are inefficient (such as using static sizes for the size or width of elements, making them non-responsive to changes in screen size).

## JS
### index.js
Gave the language navigation it's own JS, trying to make it respond to click rather than hover but its a work in progress.

## PUG
### index.pug
Added comments for clarity
### newshub.pug
Starting to toy around with it, trying to understand the logic and way it was structured so we can update the newshub. I would just re-do it from start but the code being used from wordpress and the json is far too advanced for us to re-do at the moment.

### ai.pug
Making the contents more compact so they are easier to read

### tutorials-guides.pug
Creation of page, work in progress.

## JS
### index.js
The faction containers got fixed and now work plus they have new hover and active animations.

## SASS
Small changes done to the styling of the pages, changed margins and fixed other issues.