Banners and Pages
## Content Creator Page
This update is just the creation of the content needed for the content creator page, it using a pug mixin and being pretty much "completed" for now.
## default.pug
The navbar received some re-ordering so it is ordered by longest to shortest text, giving it a cleaner look.

## Elsewhere
Not much was changed, maybe I created the sass files needed for certain pages but nothing significant was done.