1V1 (#382)
* Fixed default.pug not working
## default.pug
Small fix to default.pug so @MrRowey can continue to work on the player counter.

## index.pug
JS was on the same indentation as a block content and so it wasnt working. I indented so it works.

* 1v1 Leaderboard
## PUG
index.pug and default.pug both received updates so only the index page shows the amount of players currently online.

leaderboards.pug received some IDs in order to be able to receive the js from leaderboards.js adequately.

## JS
leaderboards.js received the JS needed to be able to pull the data from 1v1.json and make the leaderboard for 1v1. Currently, it only supports the first 100 players, soon a First, last, previous and next buttons will be added.

## Other scripts
Certain changes that were very minor were done on extractor.js, getRecentUsers.js and so on were done so the files from the API are placed in the correct folder.

* 1v1 Leaderboard
## PUG
index.pug and default.pug both received updates so only the index page shows the amount of players currently online.

leaderboards.pug received some IDs in order to be able to receive the js from leaderboards.js adequately.

## JS
leaderboards.js received the JS needed to be able to pull the data from 1v1.json and make the leaderboard for 1v1. Currently, it only supports the first 100 players, soon a First, last, previous and next buttons will be added.

## Other scripts
Certain changes that were very minor were done on extractor.js, getRecentUsers.js and so on were done so the files from the API are placed in the correct folder.

* Delete clans.json

* Delete recent.json