Leaderboard Update / Creation ## PUG Small changes to naming and the addition of certain IDs in certain elements for their use in js (for example, the global container having a #globalLeaderboard Id so it can be called in js when its clicked)
## JS
### Creation of the Pagination
The complete creation of the next, previous and switch of leaderboard type (currently only global). This was done through just removing the previous contents with the'delete' class and re-adding the new ones which then would be changed either to the next 100 players or the previous 100. Huge thanks to both @MrRowey and @Sheikah

## Other scripts
Small change done to updateLeagues.js so it moves the clans.json to the correct members folder and also adding a .gitignore to it so the json files dont get commited.