1v1 leaderboard (#383)
* Fixed default.pug not working
## default.pug
Small fix to default.pug so @MrRowey can continue to work on the player counter.

## index.pug
JS was on the same indentation as a block content and so it wasnt working. I indented so it works.

* 1v1 Leaderboard
## PUG
index.pug and default.pug both received updates so only the index page shows the amount of players currently online.

leaderboards.pug received some IDs in order to be able to receive the js from leaderboards.js adequately.

## JS
leaderboards.js received the JS needed to be able to pull the data from 1v1.json and make the leaderboard for 1v1. Currently, it only supports the first 100 players, soon a First, last, previous and next buttons will be added.

## Other scripts
Certain changes that were very minor were done on extractor.js, getRecentUsers.js and so on were done so the files from the API are placed in the correct folder.

* 1v1 Leaderboard
## PUG
index.pug and default.pug both received updates so only the index page shows the amount of players currently online.

leaderboards.pug received some IDs in order to be able to receive the js from leaderboards.js adequately.

## JS
leaderboards.js received the JS needed to be able to pull the data from 1v1.json and make the leaderboard for 1v1. Currently, it only supports the first 100 players, soon a First, last, previous and next buttons will be added.

## Other scripts
Certain changes that were very minor were done on extractor.js, getRecentUsers.js and so on were done so the files from the API are placed in the correct folder.

* Delete clans.json

* Delete recent.json

* Leaderboard Update / Creation
## PUG
Small changes to naming and the addition of certain IDs in certain elements for their use in js (for example, the global container having a #globalLeaderboard Id so it can be called in js when its clicked)

## JS
### Creation of the Pagination
The complete creation of the next, previous and switch of leaderboard type (currently only global). This was done through just removing the previous contents with the'delete' class and re-adding the new ones which then would be changed either to the next 100 players or the previous 100. Huge thanks to both @MrRowey and @Sheikah

## Other scripts
Small change done to updateLeagues.js so it moves the clans.json to the correct members folder and also adding a .gitignore to it so the json files dont get commited.