Leaderboard Update / Creation ## Styling
This update brings a lot of styling changes to the leaderboards and small tweaks to the size of the nav bar and other things.

## Functionality
Now all 4 leaderboards are operational with their own searchbar that can compare across the board. It is still a work in progress to get the missing player issue working.

## Bigger Reason
This commit is being done since my PC is having issues and I want to keep a safe copy of my work before its too late.