New frontend (#387)
* Fixed default.pug not working
## default.pug
Small fix to default.pug so @MrRowey can continue to work on the player counter.

## index.pug
JS was on the same indentation as a block content and so it wasnt working. I indented so it works.

* 1v1 Leaderboard
## PUG
index.pug and default.pug both received updates so only the index page shows the amount of players currently online.

leaderboards.pug received some IDs in order to be able to receive the js from leaderboards.js adequately.

## JS
leaderboards.js received the JS needed to be able to pull the data from 1v1.json and make the leaderboard for 1v1. Currently, it only supports the first 100 players, soon a First, last, previous and next buttons will be added.

## Other scripts
Certain changes that were very minor were done on extractor.js, getRecentUsers.js and so on were done so the files from the API are placed in the correct folder.

* 1v1 Leaderboard
## PUG
index.pug and default.pug both received updates so only the index page shows the amount of players currently online.

leaderboards.pug received some IDs in order to be able to receive the js from leaderboards.js adequately.

## JS
leaderboards.js received the JS needed to be able to pull the data from 1v1.json and make the leaderboard for 1v1. Currently, it only supports the first 100 players, soon a First, last, previous and next buttons will be added.

## Other scripts
Certain changes that were very minor were done on extractor.js, getRecentUsers.js and so on were done so the files from the API are placed in the correct folder.

* Delete clans.json

* Delete recent.json

* Leaderboard Update / Creation
## Styling

This update brings a lot of styling changes to the leaderboards and small tweaks to the size of the nav bar and other things.

## Functionality
Now all 4 leaderboards are operational with their own searchbar that can compare across the board. It is still a work in progress to get the missing player issue working.

## Bigger Reason
This commit is being done since my PC is having issues and I want to keep a safe copy of my work before its too late.