Leaderboards Completed
## SASS Changes
Deletion of unused classes, cleanup of leaderboards styling, the adding of other classes that were needed for the new dropdown buttons.

## JS Changes
The leaderboard finally is fully operational. With a time filter, search function, matchmaker filter and a clean UI that works both in desktop and mobile screen sizes.

Alot of the previous JS was trimmed and made much more compact, so the leaderboards.js are occupying as least space as needed and now can be read much more easily.

## PUG

Small changes done to certain naming classes and Ids for a cleaner and easier to read file.