Leaderboards and Extractor
## JS
Improvements made to the extractor.js file, made far more compact and using less packages and way less code repetition.

leaderboards.js also had improvements thanks to @Louvenarde's review on how to do certain processes on a more efficient manner (such as using flooring instead of slicing to get the last page of the leaderboards). Also updated to reflect the new changes done on extractor.js.

## Node-fetch
installed so we could use fetch to call API.

## AI Page
Deletion of the LOUD honorable mention due to the association's decision on the annual meeting.