Leaderboards (#389)
* Fixed default.pug not working
## default.pug
Small fix to default.pug so @MrRowey can continue to work on the player counter.

## index.pug
JS was on the same indentation as a block content and so it wasnt working. I indented so it works.

* 1v1 Leaderboard
## PUG
index.pug and default.pug both received updates so only the index page shows the amount of players currently online.

leaderboards.pug received some IDs in order to be able to receive the js from leaderboards.js adequately.

## JS
leaderboards.js received the JS needed to be able to pull the data from 1v1.json and make the leaderboard for 1v1. Currently, it only supports the first 100 players, soon a First, last, previous and next buttons will be added.

## Other scripts
Certain changes that were very minor were done on extractor.js, getRecentUsers.js and so on were done so the files from the API are placed in the correct folder.

* 1v1 Leaderboard
## PUG
index.pug and default.pug both received updates so only the index page shows the amount of players currently online.

leaderboards.pug received some IDs in order to be able to receive the js from leaderboards.js adequately.

## JS
leaderboards.js received the JS needed to be able to pull the data from 1v1.json and make the leaderboard for 1v1. Currently, it only supports the first 100 players, soon a First, last, previous and next buttons will be added.

## Other scripts
Certain changes that were very minor were done on extractor.js, getRecentUsers.js and so on were done so the files from the API are placed in the correct folder.

* Delete clans.json

* Delete recent.json

* Leaderboard Update / Creation
## Styling

This update brings a lot of styling changes to the leaderboards and small tweaks to the size of the nav bar and other things.

## Functionality
Now all 4 leaderboards are operational with their own searchbar that can compare across the board. It is still a work in progress to get the missing player issue working.

## Bigger Reason
This commit is being done since my PC is having issues and I want to keep a safe copy of my work before its too late.

* Leaderboards Completed

## SASS Changes
Deletion of unused classes, cleanup of leaderboards styling, the adding of other classes that were needed for the new dropdown buttons.

## JS Changes
The leaderboard finally is fully operational. With a time filter, search function, matchmaker filter and a clean UI that works both in desktop and mobile screen sizes.

Alot of the previous JS was trimmed and made much more compact, so the leaderboards.js are occupying as least space as needed and now can be read much more easily.

## PUG

Small changes done to certain naming classes and Ids for a cleaner and easier to read file.

* Leaderboards and Extractor

## JS
Improvements made to the extractor.js file, made far more compact and using less packages and way less code repetition.

leaderboards.js also had improvements thanks to @Louvenarde's review on how to do certain processes on a more efficient manner (such as using flooring instead of slicing to get the last page of the leaderboards). Also updated to reflect the new changes done on extractor.js.

## Node-fetch
installed so we could use fetch to call API.

## AI Page
Deletion of the LOUD honorable mention due to the association's decision on the annual meeting.

Co-authored-by: Javi Trek <96269542+FemboyJavi@users.noreply.github.com>