Navbar old (#393)
* Fix Development Environment Issues

## Fix
We commented out localization support for the time being and fixed other issues in other files due to missing package variables (such as let = require('moment').

* Navigation Bar 2.0
# Navigation
## Desktop Navbar
Received some minor updates, having the same core structure but a new fade animation added to its menus. Now both the topnavbar and the desktop navbar share the same color for a more unified structure.

The PLAY NOW button was reduced to a DOWNLOAD button.

FAF NEWS > NEWS

and the removal of other elements inside the GAME, COMMUNITY And COMPETITIVE tabs (for example, there isnt a Trainers page anymore since it will be part of the tutorials guide page).

## Mobile Nav Bar
Complete and total overhaul of the previous navigation bar. Now with far cleaner code and less JS. Occupies the whole screen, stops user from scrolling inside navbar and has a far cleaner design that syncs with the rest of the splats and the "reforged" aesthetic.

Alot of testing done on it with the inspector/dev tools of Chrome. But further testing is needed on an actual mobile device, just to make sure the font size and responsiveness is truly adequate.

## Footer
Also received a small redesign while keeping the same core structure. Now it has a much cleaner design. Updated a couple of links and terminology as well.

# Leaderboards
## Responsiveness
The leaderboard had issues with the names of usernames having the " - " character (as seen with user  "e------------") since it would break off into the next line when the username ran out of space rather than hide the overflow (increasing the height of the username column, making it not able to fit correctly with the other columns). The issue wasn't resolved directly, but instead the username column had an increase of space (at the expense of the rating column being smaller now).

Also received minor changes with the placement of the buttons for ease of use.

## Other
Smaller and trivial changes were done to certain variables or pages. Nothing worth mentioning (such as slight color changes on _variables.sass).