Client News Update
## New Features
Besides now having a much cleaner and smoother design that fits better with FAF, the new newshub comes with a couple more features.

### No more slowdowns
The previous newshub was calling the API everytime a user came (as far as I understand the outdated techniques previously used) causing users to wait about 5 seconds for the page to load (both in client and on the website).

The new client newshub caches the newshub in a json file. Meaning users dont have to wait since all the data is already loaded in the website.

### Image Sizing
Now all image sizes are supported! No need to make your images 145x97 or other cumbersome sizes. Now images are automatically re-sized.

 Images with a 16:9 ratio (2560x1440, 1920x1080, 1280x720 or anything "rectangular") will fit perfectly. Other non-rectangular images will still adjust correctly but might receive a cut-off.

 ### Unlimited Articles
 Since now the articles are on a "slider", we dont have a limit with how many articles we can have without sacrificing design. So 5, 7 or 12 articles will all display equally well!

 ### Responsiveness
 This design has media queries and responsive unit sizes (such as vw) to always adjust to the users preferences. Therefore, the news will display adequately whether it's on a full screen window or 1/4 of the monitor screen.

 ### More Information Less Clutter
 The new design also provides more links for the user. With buttons for joining the discord,  twitch and youtube FAF content being one click away, plus new sections for new players, community tools/spaces and development, plus a Support FAF button and a Report a Bug button all in one compact and clean design! Now Users can have an easier time accessing FAF tools and spaces from the client News!

 ## Other
 Many of the images inside the images folder have been moved to a logos folder in order to maintain a better order inside the folder (rather than having 40 images on one folder, have them on their own folders).