Express JS cleanup
## Tournament Page
The tournament page has received a small cleanup and work has started on it. Currently It has stopped since I'm learning how to go through oauth2 for challonge and using passport js/express js.

## Express.js
Received a small cleanup in code (taking out repetiion and putting it into a for loop/for each statement). Will continue to receive updates and cleanups as I continue to learn more about it.