News Articles Update
## News articles now have their own page/url
Yes you read it right, now the FAF website has the capability to not only pull the articles from the wordpress and render them correctly. But, it also is able to create urls for users to share a specific article!

## UI Update
The newshub page has also received some tweaks on the graphical side.