File Update ## Deletion of Files Certain files were deleted either because they were re-incorporated somewhere else or because they are not currently in use.
This is being done to reduce the amount of noise in the backend code since I'm quite inexperienced with the backend and the less files the easier it will be for me to understand the system.

### extractor.js
Changes were done to certain naming conventions for better readability, plus some comments were added.

### getRecentUsers.js
Changed into simpler/easier to read code. It has its module on express.js to run every 30 seconds. It takes the recent ranked (with at least 1 ranked game) players of the last 4 hours.