Pug Tags Bug Fix ## Deletion of Files Certain files were deleted either because they were re-incorporated somewhere else or because they are not currently in use.
## navigation.js / Navbar has its own js file and doesn't share it with index page
So before, the home page and navbar shared the same js file. This has been fixed so only the navbar only brings the js it needs instead of every page having the navbar and the index/home page js. This also caused some console errors on non-index pages since it wouldnt be able to find elements that are only on the index page.

## getRecentUsers.js
Stopped it from console logging since it would do so every 30 seconds...

## 404.pug 500.pug
Both now show a banner image with the same message.