.env variables and logging in and out # Backend Changes
## .env
Some variables that should have been secret have been put into the .env document to stop any funny business.

## login in and out
Now the website is able to verify if you are logged in (through discord) and able to make only select routes checked if the user is logged in or not (currently all for testing purposes).