Flash Message + Play Page
### Flash Messages
Now the website features flash messages! They can feature custom colors and have the capability of being shown in only specific pages. This matters because we dont want to place a "Dont register with a gmail account" flash message on every page but maybe only on a couple of pages such as /accounts/register for example.

### Play Page
The play page has received a couple of fixes gramatically, added more resources to improve the process and a couple of styling changes thanks to the feedback received in the forum post.