Undoing updates to backend
### Reverting to original
Due to a lack of testing capabilities. All iniatives to update the backend have been been undone since the only available testing enviroment is the test server but its not good to rebuild it 100 times everytime I change a line of code. Ergo, will try to see how much can work with the outdated "request" library that is currently being in use"