Clan Management + Register
## Clan Management

I'm currently trying to improve the invitation process to become easier (copying the inviting link to the user after they create the invitation).

## Clan Pages
Have had an increased setTimeout (from 100ms to 1000ms) to see if that works in the test server.

## Register
Received minor changes in styling and HTML placing to see if it fixes the register button.