Competitive = Leaderboards
### Tournaments = Leaderboards
Tournaments page will be cut out since there is a lack of a work force/maintainer to update it with the latest tournament. Maybe in the future it could be re-inplemented but I want to avoid having a page whose maintanability is unsure when the newshub is already having issues finding a maintainer.

### Mobile Navigation
Received a small fix due to some color changes in variables.sass that dont translate to javascript.