Clan Member Leave
### Leaving a clan
Its a bit of a shoddy solution but when members click on "My clan" and they arent leaders, they'll be greeted by a "Leave my clan" option. However, this only appears if they try to get to clan/manage rather than just going to their clan. But the average user wont notice the issue.

### Minor Styling Changes
Done to home page.