Minor Styling Changes
### .containerCenter
before, alot of protected routes had the .container class, this class is also shared with the markdown files. However, I wanted markdown to have a text align left and account/clans forms to be align center. So a change had to be done from .container to .containerCenter

### Testing Clans
Some other minor changes were done to getClans.js in order to test with clans from the test server.