Testing errors + Clans
### getClans
Now takes away either just the ? or ?m. before it was taking out any m or ? it would find. Making clans like FEM be searched as FE.

### Extractor -> axios
All fetch requests have been changed to axios just to see if it will work with the current node image of the test server which doesn't support npm 9.1.1 and therefore crash the server when the request fails.