Report and Invitation Page
### Reports
The report page has been updated and re-accomodated to the new frontend. It has been tested and should work correctly on the test server.

### Invitation
Inviting someone to a clan resulted in them being redirected to the wrong url. This has now been fixed.

### Navigation Images
The social media icons used before with font awesome are now downloaded svgs.