Leaderboard Style
### Front-End
The style of the leaderboards, its buttons and its organization has changed.

### Back-end
Leaderboards is now organized from first player to last player. Before it was sorted out backwards so the #1 player was actually the last player.

There was an attempt to create sorting for winrate and total games but they will need more work.