Steam Link, Navigation, lobby_api (#428)
### SteamLink
Steamlinking routes have been all transformed back to their outdated routes for the sake of a hotfix in order for steamlink to work properly.

### Navigation
Received small superficial update to be better looking and give users a better experience

### lobby_api
reinstated after it seems it is actually being used and is not some useless piece of code. It provides users with a mirror of the users, their maps and lobbies in the FAF client.