Mobile Navigation Update (#431)
* Mobile Navigation Update

### Cleanup
The previous mobile nav bar was a bit messy and done amateurly. Not much changes in this update but now naming conventions, usage of elements and amount of code are more compact, cleaneer and easier to read.

### Better Menu
Some options in the navbar such as Game or Community open up a menu to other links. These have received an update so the navbar doesnt break and the user has a better experience when navigating the website through the mobile.

### Mobile Login!
Finally the website is able to allow users to login and have a login navigation menu through mobile. And the account pages have also become more responsive to suit smaller devices.

### Other
There was a small text correction on the play page (create your account -> create and activate your account) and the leaderboards having a minor visual touchup in font size.

* Remove Console Log