Account Settings
### Bye Bye separate account pages!
Before, changeEmail, Password and Username all had their own page. Now, all three (and more to come) can now be set on an account settings page. Which looks more cleaner and helps the navigation bars as well to not have so much clutter.

This also helps cut out 3 additional pug pages, making the website more compact.

### Request out, Axios in
Request is an outdated and unmaintained dependency that one day will stop working. I'm cleaning the back-end code to be more readable and use Axios, a maintained and more modern HTTP requests.

### Remember URL if not logged
Now when an non-logged user tries to access a protected route, they should be redirected to that url after login. However, if two users login around the same time, it is very probable the first one will be redirected to the URL of the second one. This needs more testing but its less frustrating than always being redirected to the home page.

### Lobby API
Received a change on its env variable so it migth finally work