Page Change + Validation Errors
### Account Settings
Now when change Password, Email or Username is requested, the user gets sent to that specific setting instead of always defaulting to username settings.

### Error Handling / Flash Error
It seems the website wasn't really doing alot of error handling when it came to validating data. For example, the website wouldn't handle the error of a new password not matching it being confirmed or if it was actually longer than 6 characters.

Therefore, the way validation is currently set needs to be re-done in order to make sure it works correctly.