Route Refactoring
### play.sass
Received minor changes to image width.

### accountRoutes.js
Its unnecessary to have a middleware for validating forms. I've learned a way to get the form into an array so we can check the form with express-validator BEFORE we do the req, res function. Otherwise, if we check after we call the function, express-validator doesn't work.

### error.js errorChecking
Now error management has its own function in order to avoid alot of code repetition. It also will help with having to do a change in one function rather than in dozens of pages.

### _mixins.sass _navigation.sass
Both received small changes that will stop the website from being so "jumpy" everytime we move from one page to the other. It was simply due to a transition effect being applied to the main containers.

### Account + Clan Refactored
Both account and clan routes are being refactored to improve the readability and long term maintainability of the website. Making them shorter and simpler will ensure they are easier to change.