Clan Route Refactoring
### Axios in, request out
Most of the updated clan routes are just switching request for axios and cleaning up alot of stuff.

There is some issues where flash messages are harder to display but to be honest, a lot of the times the user doesn't need a flash message to tell them "You've been accepted in x clan", being redirected to the clan itself seems like the simpler option for now.

### flash message HTML
Now flash has flash.hasHTML, which is just like flash.messages but accepts/renders HTML. Used currently in the clan link.

### settings.sass / settings.pug
Received minor changes to center buttons and to give them more space.